Homework number five solved

+++ ObjectOrientedDesign/TestDocCli/Program.cs
@@ -20,7 +20,7 @@ catch (KnownUserErrorException knownUserError)
 {
   var reporter = new ErrorReporter();
   reporter.Write("ERROR", knownUserError.Message);
-  Environment.ExitCode = knownUserError.ExitCode;
+  Environment.ExitCode = (int)knownUserError.ExitCode;
 }
 catch (Exception exception)
 {
@@ -33,4 +33,4 @@ catch (Exception exception)
 // WE have a IConsole where we can define the console operations for write and read
 //  - The read opeartion might complex so, we created a different class that will implement IConsole (or not) to perform the read operation
 
-// TODO: [UPDATE] Create the diagram of this project v2.
+// SOLVED: [UPDATE] Create the diagram of this project v2. It is saved on the files Updated-Diagram-Project-v2.drawio and Updated-Diagram-Project-v2.drawio.png
